strip numpy from engine
