+++ backend/requirements.txt
@@ -8,10 +8,6 @@ requests==2.31.0
 python-multipart==0.0.6
 python-jose[cryptography]==3.3.0
 passlib[bcrypt]==1.7.4
-torch
-transformers
-sentence-transformers
-numpy
 aiofiles==23.2.0
 httpx==0.25.2
 schedule==1.2.0
